Added working randomization of battleship board
Also added logic to add to the players score. As well as a screen declaring the winner when the last boat is sunk. Added an option called grid size to the main screen which replaces settings and allows the player to define the dimensions of the grid, hasn't been integrated to the other classes however.